fastapi/requirements.txt
ipython
fastapi
uvicorn[standard]
paho-mqtt
sqlalchemy
psycopg2-binary
python-dotenv
shapely
geoalchemy2
# -e ../../bdoh-core
